Show real source code under every example
Each example is now a standalone file under src/examples/, and App.tsx
imports both the component and its source via Vite ?raw imports, so the
displayed code can never drift from what runs. A shared ExampleSection
renders title, description, the live demo, and a collapsible code block.

- planet examples split into their own files (demo-only, page chrome
  moved to App)
- react-example gains a typecheck script (tsconfig needed types +
  ES2022 lib fixes that were previously unchecked)
- e2e test asserts every section has a working Show code block
