+++ packages/react-example/package.json
@@ -6,6 +6,7 @@
 	"scripts": {
 		"dev": "vite",
 		"build": "vite build",
+		"typecheck": "tsc --noEmit",
 		"preview": "vite preview"
 	},
 	"dependencies": {
