+++ packages/react-example/tsconfig.json
@@ -2,9 +2,10 @@
 	"compilerOptions": {
 		"target": "ES2020",
 		"useDefineForClassFields": true,
-		"lib": ["ES2020", "DOM", "DOM.Iterable"],
+		"lib": ["ES2022", "DOM", "DOM.Iterable"],
 		"module": "ESNext",
 		"skipLibCheck": true,
+		"types": ["vite/client", "node"],
 
 		/* Bundler mode */
 		"moduleResolution": "bundler",
